LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/flow-content-0/dialog-display-expected.txt

CONSOLE MESSAGE: line 3463: Error: assert_equals: expected "block" but got "inline"

Harness Error (FAIL), message = Error: assert_equals: expected "block" but got "inline"
